+++ gpiozero/exc.py
@@ -1,0 +1,19 @@
+from __future__ import (
+    unicode_literals,
+    print_function,
+    absolute_import,
+    division,
+)
+
+class GPIODeviceError(Exception):
+    pass
+
+class GPIODeviceClosed(GPIODeviceError):
+    pass
+
+class InputDeviceError(GPIODeviceError):
+    pass
+
+class OutputDeviceError(GPIODeviceError):
+    pass
+
